+++ SampleProject/Assets/Tests/EditMode/LastBeaconStoreTests.cs
@@ -13,7 +13,9 @@ namespace Persistly.Unity.LastBeacon.Tests
 
             var profile = new LastBeaconProfile
             {
-                SaveId = "sv_01HXYZ",
+                ProfileSaveId = "sv_profile",
+                ProfileSessionToken = "pst_profile_session",
+                CharacterSaveId = "sv_01HXYZ",
                 Version = 7,
                 Config = new LastBeaconConfig
                 {
@@ -38,7 +40,9 @@ namespace Persistly.Unity.LastBeacon.Tests
             store.Save(profile);
             var reloaded = store.Load();
 
-            Assert.That(reloaded.SaveId, Is.EqualTo("sv_01HXYZ"));
+            Assert.That(reloaded.ProfileSaveId, Is.EqualTo("sv_profile"));
+            Assert.That(reloaded.ProfileSessionToken, Is.EqualTo("pst_profile_session"));
+            Assert.That(reloaded.CharacterSaveId, Is.EqualTo("sv_01HXYZ"));
             Assert.That(reloaded.Version, Is.EqualTo(7));
             Assert.That(reloaded.Config.CharacterName, Is.EqualTo("Ayla"));
             Assert.That(reloaded.State.Workers, Is.EqualTo(4));